Basic functionality
Adding basic functionalities for the player and enemies to play the game.  Player starts off with 3 "lives", as in it can die upto 3 times.  When the 3rd life dies, it's game over.

Once the player crosses over to the water, then it win and it'll re-position itself to the starting point.